# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Melete",
    "version": "0.1.0-pre",
    "summary": "An open-source, self-hosted, model-agnostic personal assistant.",
    "description": "Give Melete a responsibility, close the tab, come back to progress, a result, or one precise question. This document describes the v0.1 HTTP surface. Nothing here is stable yet: the release is pre-release and endpoints may change until v0.1.0 is tagged.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Default self-hosted address"
    }
  ],
...
  "paths": {
...
    "/messages/{messageId}/reactions": {
      "post": {
        "tags": [
          "reactions"
        ],
        "summary": "React to a message with one emoji",
        "description": "A message is an event, and its id is that event seq. The reaction is persisted and streamed like any other event, and a client draws it on the message bubble rather than as a row of its own. A thumbs-down from a person counts the result it lands on as two unread ones; a thumbs-up clears the unread streak. Reacting twice with the same emoji records one reaction.",
        "parameters": [
          {
            "in": "path",
            "name": "messageId",
            "schema": {
              "type": "string",
              "description": "Message id: the event seq"
            },
            "required": true,
            "description": "Message id: the event seq"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "emoji": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 24
                  }
                },
                "required": [
                  "emoji"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recorded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "reaction": {
                      "$ref": "#/components/schemas/__schema98"
                    }
                  },
                  "required": [
                    "reaction"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "No such message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/__schema58"
                }
              }
            }
          },
          "409": {
            "description": "That event is not a message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/__schema58"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "reactions"
        ],
        "summary": "The reactions drawn on one message",
        "parameters": [
          {
            "in": "path",
            "name": "messageId",
            "schema": {
              "type": "string",
              "description": "Message id: the event seq"
            },
            "required": true,
            "description": "Message id: the event seq"
          }
        ],
        "responses": {
          "200": {
            "description": "Reactions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/__schema99"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{jobId}/reactions": {
      "get": {
        "tags": [
          "reactions"
        ],
        "summary": "Every reaction on one job stream, for a client rendering a transcript",
        "parameters": [
          {
            "in": "path",
            "name": "jobId",
            "schema": {
              "type": "string",
              "description": "Job id"
            },
            "required": true,
            "description": "Job id"
          }
        ],
        "responses": {
          "200": {
            "description": "Reactions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/__schema99"
                }
              }
            }
          }
        }
      }
    },
...
    "/artifacts/{id}/content": {
      "get": {
        "tags": [
          "artifacts"
        ],
        "summary": "Retrieve an artifact in the authenticated space",
        "description": "Returns the recorded bytes only while their hash matches the artifact receipt. Audio can be played directly; a single byte range can be requested for seeking.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "description": "Artifact id from the action receipt"
            },
            "required": true,
            "description": "Artifact id from the action receipt"
          },
          {
            "in": "header",
            "name": "Range",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Artifact bytes",
            "content": {
              "audio/wav": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "206": {
            "description": "Requested byte range",
            "content": {
              "audio/wav": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "A session is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/__schema58"
                }
              }
            }
          },
          "404": {
            "description": "No matching artifact in this space",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/__schema58"
                }
              }
            }
          },
          "416": {
            "description": "Requested range is outside the artifact"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "__schema41": {
        "type": "string",
        "format": "date-time",
        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
      },
...
      "__schema57": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "detail": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          }
        },
        "required": [
          "code",
          "message"
        ],
        "additionalProperties": false
      },
      "__schema58": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/__schema57"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "__schema98": {
        "type": "object",
        "properties": {
          "message_id": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,17}$"
          },
          "emoji": {
            "type": "string",
            "minLength": 1,
            "maxLength": 24
          },
          "by": {
            "type": "string",
            "enum": [
              "person",
              "assistant"
            ]
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^job_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
              },
              {
                "type": "null"
              }
            ]
          },
          "seq": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "created_at": {
            "$ref": "#/components/schemas/__schema41"
          }
        },
        "required": [
          "message_id",
          "emoji",
          "by",
          "job_id",
          "seq",
          "created_at"
        ],
        "additionalProperties": false
      },
      "__schema99": {
        "type": "object",
        "properties": {
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/__schema98"
            }
          }
        },
        "required": [
          "reactions"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "melete_session"
      }
    }
  }
}